Update README to record the death of this utility
*************** THIS PROGRAM IS NO LONGER MAINTAINED BY X.ORG ***************

xsetmode sets the mode of an XInput device to either absolute or relative.

xsetmode still works on devices that support multiple mode, the hooks are
still there so functionally there's nothing wrong with it, but it is no
longer actively maintained as the xinput program has replaced it, with:
  xinput --set-mode "device name" ABSOLUTE

Since xinput is actively maintained, X.Org recommends using that instead.

*****************************************************************************

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>